requirements.txt
duolingo-api==0.3
plotly==2.0.1
numpy==1.13.1
